+++ src/main/java/nextstep/ladder/domain/ladder/Ladder.java
@@ -1,4 +1,6 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.ladder;
+
+import nextstep.ladder.domain.player.Player;
 
 import java.util.Collection;
 import java.util.List;
