+++ src/main/java/nextstep/ladder/domain/player/Player.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.player;
 
 import java.util.Objects;
