+++ src/main/java/nextstep/ladder/view/output/ResultView.java
@@ -1,8 +1,8 @@
 package nextstep.ladder.view.output;
 
 import nextstep.ladder.Visible;
-import nextstep.ladder.domain.Ladder;
-import nextstep.ladder.domain.MatchResult;
+import nextstep.ladder.domain.ladder.Ladder;
+import nextstep.ladder.domain.result.MatchResult;
 
 import java.util.Scanner;
 import java.util.stream.Collectors;
