+++ src/test/java/nextstep/ladder/domain/GameOrganizerTest.java
@@ -1,5 +1,12 @@
 package nextstep.ladder.domain;
 
+import nextstep.ladder.GameOrganizer;
+import nextstep.ladder.domain.ladder.Ladder;
+import nextstep.ladder.domain.ladder.LadderLine;
+import nextstep.ladder.domain.player.Player;
+import nextstep.ladder.domain.player.PlayerGroup;
+import nextstep.ladder.domain.result.LadderResultGroup;
+import nextstep.ladder.domain.result.MatchResult;
 import nextstep.ladder.dto.PlayerResult;
 import org.junit.jupiter.api.Test;
