+++ src/test/java/nextstep/ladder/domain/player/PlayerGroupTest.java
@@ -1,5 +1,7 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.player;
 
+import nextstep.ladder.domain.ladder.Ladder;
+import nextstep.ladder.domain.ladder.LadderLine;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
