+++ src/test/java/nextstep/ladder/domain/player/PlayerTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.player;
 
 import org.junit.jupiter.api.Test;
