+++ src/test/java/nextstep/ladder/domain/result/LadderResultGroupTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.result;
 
 import org.junit.jupiter.api.Test;
